Detect duplicate tasks
Added a constraint that does not allow tasks that has been added to be
added again into the Task List. Tasks with different types (Event, Todo,
and Deadline) with the same description will be considered a different
task.

Two deadline tasks will also be considered different if their deadlines
do not end at the same time.