Used Java assertions to define assumptions about the program state.
Added assertions in Parser class to define the assumption that empty
inputs will not be passed to the functions in Parser class to be parsed.

Added assertions in TaskList class to define the assumption that the
task number has to be a valid array index, more than or equals to 0.

Added assertions in Task, Event and Deadline classes to define the
assumption that the task number has to be more than or equals to 1.