Improved code quality.
Improved readability by deleting dead code and redundant variables.

Made the happy path prominent by using guard clauses at the top of the
function and throwing an exception if it enters the guard clause.

Minimized code duplication and edited my comments to be in the format of
writing to a reader.