Update Commands enumeration with specialized help messages
Instructions provided on how to use each command (shown
when user types the 'help' command)  is currently statically
typed and printed in the parseAndExecuteCommand()
method's switch.

Placing the instructions for each of the possible commands in the
Commands enumeration makes it easier to maintain and add to.

Let's make sure to update the enumeration each time we add
a new command instead of cluttering the parseAndExecuteCommand()
method