Add duplicate check
The check is needed because the user may add duplicate task to the list.

The program will now check whether the task to be added is already in the list.
If it is in the list, the task will not be added.

The duplicate task will not be added because it has the same details which makes
two same instances of task are redundant and may confuse the user.