Throw exception if the save file is in the wrong String format
The GUI would not load at all as to prevent the user from using the app.
If the app were to be used, how should the save file be loaded? Using a
new save file would delete all the data (which is undesirable), while
finding an exception at the start of the save file would also require
deleting all the data (anything after the problem). Thus, not loading
the GUI tells the user about the problem with save file and gives them
a chance to fix the save file.