Implement task update functionality
Implement ability to update tasks

Currently tasks cannot be edited once created

This limits functionality of the bot as editing
tasks is a natural thing to do.

Including and handling an update keyword
allows people to use this function.

The keyword is necessary to inform that this is
a new function. It also enables seamless
incorporation along with the existing commands

Update format is:
update [task index] [new name] [/param] [new param]
If a parameter or the name is not being updated,
you may use the '*' to notify the program that
you wish to keep the existing value.

Consequently, '*' is now a reserved character and cannot
be used like a typical one. This decision limits inputs
slightly but the decision was made because it is unlikely
that a user will enter just a '*' as a name or description
to any task.

Examples of update commands:
 - update 1 eat /by 2020-11-11
 - update 4 say hello
 - update 2 * /from * /to tomorrow