Add assertions
DeleteCommand and MarkCommand should have index > 0. Storage load method should also have file (create file if it does not exist).

Assertion of index > 0 ensures that the condition is always satisfied when assertion is turned on.
Assertion of file.exists() ensures that a file always exists after being created or read.