Implement Assertions for validation
The code does not have any internal validation.
This may allow errors during development that
may go unnotice, and may make the code unreliable.

This may make the code unreliable as sometimes
external tests such as JUnit testing, not not
be able to create all possible scenarios that
may happen during production.

Add assertions to classes with in the codebase.

This will allow me to internally test our codebase
before going to to run JUnit for external testing.
This allows my testing to be comprehensive and to
catch more unexpected errors that we may not have
thought about during my development.