Ensure code quality in codebase
The code base has duplicated code, unclear naming
conventions. This makes it unnecessarily long and
complex.

This makes the code base hard to maintain and edit
in future iterations, and harder for other
developers to read and implement their logic as
well.

Refractor and fix naming conventions across the
java files. Additionally, removed repeated code
from the main UI java files to shorten the lines
of code.

This allows for better readability
for other reads, making the code more concise and
easier to understand as well.