Implement undo command
User could not undo the recent commands they made in the application.

Create a new class named Undoable that follows the Memento pattern.
This class should contain the previous state and have an undo
function. Any commands eligible for undoing should implement this
undo function. Additionally, create a StateManager class to manage
history stack for the undo feature.

The code tries to follow the design pattern for such feature providing
a clear structure for managing undoable commands. Though it can be
improved in the next iteration with max stack size and not deep
copying the state for each execute and undo commands that occurs.