Ensure code quality and fix errors in codebase.
The codebase has some naming style warnings, and
bad variables names such as c or n.

These do not provide enough detail for the other
readers, which can cause it to be misleading.

Edit the naming violations to ensure that they
are compliant with the convention.

This will increase code reability and allow
my readers to have a better understanding of
the code.