Use assertions to document invariants
Various parts of the codes currently run on assumption that it will
hold true. (e.g. int statusValue will always either be 0 or 1)

This could result in major bugs down the line if the assumption
made is wrong.

Using assertions in these places will ensure that the assumptions
being made are actually true.

This will help us in debugging as it will allow us to see if any of
our previous assumptions fail.