Refactor long method in Storage class to shorter method
The parseInput() method is almost 50 lines long which makes it
slightly harder to read and understand.

By separating the logic of every case statement to one method each,
the parseInput() method would be much shorter.

By separating the logic of every case statement, we would also be
following the SLAP principle. This would also overall improve the
readability of the code.