Add assertion to create() method
The create() method has assumptions about the length
of the split input string which are true after a series of checks.
These are however not very explicit and may break if some checks
are rewritten.

Adding asserts will make these conditions more explicit to the developer
as well as ensure they are maintained.