Add 'help' feature using streams to iterate over enumeration of commands
Help feature is implemented by printing a list of statically typed
instructions for each of the commands listed in the Commands enumeration

Each time a new command type is added, it is added to the enumeration
along with its 'help message' (instructions for how to use the command)

Compilation of the help message requires you to obtain the string by
calling command.getHelpMessage() for each of the commands

Using streams to iterate over the enumeration to generate the list is
much more efficient.

Let's use streams where possible to reduce the amount of hard coding.