Add Cat C-MassOps feature
Add a new feature where users can perform some mass operations on
all their current tasks.

This makes it easier for the user to do a mass operation and the
user do not have to manually say mark each task as done at
the end of the day, but rather can just do 1 mass operation to
mark all the tasks as done.

Let's,
* Update DeleteCommand.java to support mass delete of all tasks
currently in the TaskList.
* Update MarkCommand.java to support mass mark of all tasks
currently in the TaskList.
* Update UnmarkCommand.java to support mass unmark of all tasks
currently in the TaskList.