Utilize Java Enums for different types
The ActionTask uses strings to represent different
types of actions. This allows it to handle a fixed
set of variables.

This could lead to errors and makes the code hard
to implement, as developers have to read and match
the words for the corresponding actions.

Let's change the String list into an Enum, and use
it to indicate the respective action type of
ActionTask.

This will allow only defined types to be created,
enforcing proper type security and ensuring that
it is easy to define and debug.