Add java assertions to commands
As the program is designed to handle most of the misinputs, the only place java assertions make sense is in the commands.

Executing the commands while passing an uninitialized taskList is a sign that the program is already in a broken state. Thus assertions is used instead of catching for errors.