Add sorting to Storage
The code does not sort Tasks according to their
deadlines or start date.

This may lead users to miss their deadlines as
they are not chronologically ordered.

Implement comparator to each of the Tasks's
subclass.

This allows us to sort the Tasks according to
their urgency, if applicable, and name.