Add response classes and UI
Right now, as the week 6 better gui incentive outlined, the chat like Ui
is not really that indicative of the interaction between the app and the
user.

Therefore, I'm refactoring some of the code to make the UI emit a
Response class for the app, and this respose will seem very different
from the user's input.