Added an extension C-Update.
With this new extension, users can edit their task details by
running a command in this format:
update <taskNum> /<desc/from/to/by> <newInfo>

Added methods in Parser class to parse this new command, as well as
methods in TaskList, Storage, Task, Deadline and Event classes to
support this new extension.

If the user tries to update a non-existent attribute for a task
(e.g updating a /by attribute for a todo task), a DukeException
will be thrown and printed in the DialogBox.