Improve task storage printing format
Tasks are currently not printed initially when program is re-run, making
it difficult for users to access their todo list. This change is
necessary to enhance user experience and facilitate the use of the app.

Refactor the task printing template to be clearer. The current format
lacks clarity and doesn't display the todo list when the program is
re-run. The revised printing template aims to address these
shortcomings.