Update application mainClass to duke.Launcher
The mainClass in build.gradle was updated from duke.Duke to duke.Launcher.
This change was necessary because the application is now a JavaFX app,
which needs to be launched from the duke.Launcher class.

Updating the mainClass ensures that the application starts correctly when
built using Gradle. This change does not affect the functionality of the
application, but it is crucial for the build process.