BCD-Extension B-ViewSchedules - implemented the feature
The current program lacks a feature to view tasks in the form of a
schedule.

It's essential to implement it as it allows users to view their tasks
and plan for them more effectively.

This is an initial implementation of the View Schedules feature to allow
users to view their deadlines and events on a calendar format.

This is done as deadlines and events consists of dates and times, so it
is easier for users to view the tasks in a calendar format.

The current version may be limited in terms of customization and
functionality, and further development is planned to expand on the
feature, provided there is enough time to do so.