The chat bot currently does not support displaying the list of events happening on a particular day.
Displaying the schedule for a specified date will be useful for most users.
The added convenience allows them to plan their activities for the day and
hence, more room for productivity and organisation.

Users can execute the function by entering "schedule <date>" where the date is in
yyyy-mm-dd format, e.g. schedule 2024-03-01.