Add new Exception classes
There is no exception to detect if the user is trying to mark a task
that is completed and unmark a task that is not done.

Let's handle these exceptions by adding a new exception and checking the
status of the task before letting the user mark it as done or undone.

Also add bug fixes for the GUI (The name "Snom" is now correctly
displayed.

Todo: Remove the dead code from the remaining files.