Add Parser.java
Additional class and functions aid in parsing of user input.

The parse command returns a command of a specific type according to the user input.

Let's abstract the if-else statements and simplify them in Parser.java.

Using abstraction allows simplification and encapsulation of parsing and returning of the different commands.