Remove dead code, abstract each switch case into a method
The methods used to abstract switch cases are private as they are only
used in the class itself. This helps to keep each method short so that
it is easier to trace through them. To reduce repetition, some methods
may not be fully suitable for some cases, but are used as long as they
are similar enough.