Add assertions to note important assumptions
In Whisper class, assertion is added to check validity of assumptions
regarding the initialisation of Ui, Storage, and TaskList.

In the Parser class, included assertions to verify assumptions about the
input parsing logic and command creation.

In the Command interface and implementing classes, used assertion to
ensure that critical conditions are met during command execution.

These assertions serve as documentation for developers and help in
identifying potential issues during development or maintenance.