Refactor code to utilise Java Streams in ListCommand
In ListCommand class, utilise `forEach` with a method reference to
iterate over tasks and print each task using the `printTask` method of
the `Ui` class.

The changes leverage Java Streams to make the code more concise and
potentially improve readability. The modifications maintain the
functionality of the commands while adopting a more modern and
expressive coding style.