Refactor code to improve quality
There are some instances of poor coding practices displayed in previous
commits which hinder understanding of code, making debugging harder.

Let's,
* Modify classes (AddCommand, AddDeadlineCommand, AddEventCommand,
AddToDoCommand) to extract out common functionalities of the three
subclasses of AddCommand into a method addAndStore() to better
adhere to SLAP and reduce code duplication

* Extract out different sections of the parse() method from the Parser
class and the readAndAddTask() from the Storage class to reduce the
length of the method body, while also adhering to SLAP better by
providing an additional layer of abstraction.

* Re-implement the extracted segments from parse() into methods
(parseDeadline(), parseEvent()) using Java Streams to reduce deep
nesting of code logic

* Create local variables for min/max values of (minutes/hours) for
twenty-four hour and Am/Pm time formats respectively to reduce the use
of magic numbers

* Remove unnecessary Function calls, loops, try-catch blocks to
simplify the code and make it easier for others to understand it