Fix some coding style violations that were listed by the CheckStyle in Gradle.
Fixing these coding style violations should improve readability and ensure that the code is using the same coding style for this project.

Should anyone else make any changes to this project and use the same coding style, it will make it seem that this project has been done by a single person, thus improving the code quality of this project.