Configure build.gradle file
The main class has changed, and it will affect the referencing on the
variable mainClassName and archiveFileName.

Changing the value of mainClassName to Launcher and the value of
archiveFileName to 'chronos.jar' allows us to create a fat JAR file
containing the JavaFX libraries. This is essential for the GUI
implemented for the Chronos application.