Add Java's assert feature
It serves as a tool to document important assumptions that should hold
at various points of the code.

Adding assertions in Parser.java helps to check whether the task list
is not empty upon adding a todo, deadline or event, as well as, to check
whether the retrieved task is not null using the getTask() method.