Improve code quality
Shifted the functionality of the commands out of the Destiny class
and into individual command classes.

New classes created:
* Command (abstract)
* DeadlineCmd
* DeleteCmd
* EventCmd
* FindCmd
* ListCmd
* MarkDoneCmd
* MarkNotDoneCmd
* ToDoCmd

Changes have also been made to the parser class to accommodate the
shift in functionality.

All this was done to shorten the length of the getResponse method in
Destiny because it was too long.