Abstract out command input word into respective Command subclasses
The Parser checks if the user input matches certain keywords and
calls the corresponding command classes. This is hard-coded in the
parser.

Abstracting this command keywords out to the respective Command
subclasses makes sense, reducing potential human error.

Let's put the keywords in the Command subclasses, and have the
Parser access them.