Bug fix: resolve index out of bound error in TaskList's toString
Previous implementation will throw error when listing the taskList
if the taskList is empty.

Let's add condition to check if list is empty or not before
accessing the last task in the list, and inform user the list is
currently empty if it is.

This change will resolve the error and inform user to add tasks.