Improve Javadoc comment for the view method in `TaskList`
The Javadoc comment for view was not descriptive enough.

Editing the Javadoc comment to be more descriptive is a small step in
improving the overall code quality of the project. This will be
beneficial for future developments.

Let us change 'Filters' to 'Returns' and improve the grammar of the
sentence.

It is customary for the Javadoc comment of return statements to start
with the word 'Returns', a verb.