Add tag command
Users are unable to tag their tasks. This can be troubling when
keeping track of similar tasks.

Provide a way to tag items e.g., tag a task as #fun.
Tags can be filtered and displayed in future functionality.

Tags are implemented as a string and only 1 tag is allowed per task.
Tags can be identified with a #hash beside the task name.