Add User UserGuide
Current code base has no UserGuide created, on
GitHub under docs.

This may cause users to use the application
without clear instructions, thus not allowing
them to effectively use the features of BobBot.

Add comprehensive user guide to the docs folder,
for GitHub and users to refer to.

This will allow users to easily refer to the guide
created by the sole developer of BobBot, allowing
them to effectively use the application.

Referred to the links provided for this branch.