Edit auto-save Did not implement save properly before.
I had 2 different lists (one for the current iteration of tasks, and one that was saved in my file), but I could not edit (mark, unmark, delete) the ones that were in the file. Only my current task list was editable.

I made the two list one and the same using one ArrayList tasklist for all tasks.