Storage class: load the tasks from the list.txt
Storage class have some unintended mistake while loading by not recognized done task.

Trimming the status from the status collected from the list.txt to prevent this issue. For example, trim helps to get rid of the extra whitespace that causes the current error.

Using trim is suitable as it conveniently removes unnecssary whitespace in the status.

Refer to this website on how trim is used
https://www.w3schools.com/java/ref_string_trim.asp