Duke classes: incorporate JavaFX GUI
Duke ChatBot uses an interactive and boring textUI which makes in less
compelling for uses to want to use it.

Lets use JavaFX to make the GUI of Duke to be more interactive and add
the much needed vibrancy to compel users to use Duke.
Lets add a viewport for users to see their messages and the replies
returned by duke.
Allow users to interact with duk with a text area and a button to send
their inputs

JavaFX is used as it offers much support with Java which Duke is build
on. Furthermore, in the context of 2103T much support is provided by
the previous batch that used t.

Command Classes: Extracted each Command to have their own classes

The behaviors of the different command is inline with the execution
of Duke run() method.

While the commands have similar behaviors, each command still have
their different purpose. Therefore, code refactoring is done to
seperate them into classes to achieve:
        1. SLAP HARD
        2. DRY(avoiding code duplication)
        3. Scalability

Created an abstract Command class and have the different commands
inherit from the command class and to define their own behavior
in the execute method

Using inheritance allow for the methods in Duke to be more flexible
and use polymorphism to have their own behaviors