Update logic flow and response
MainWindow class: handle TerminateException

With the implementation of a GUI, MainWindow class
now handles the running and termination of the application.
As such, the "bye" command is no longer able to close the
application.

MainWindow catch and handle TerminateException.

This allows MainWindow to know if the application
needs to be terminated.