Add assertions
Assertions are missing from class methods.

Some assumptions need to be true at the start and end of an execution to
ensure correctness.

Adding assertions allow for better runtime verification of these
assumptions. For example, in Ui class, at the end of printList()
function, an assertion is added to ensure the counter for tasks in the
list did not decrease.