Add AddCommand.java, ByeCommand.java, Command.java, DeleteCommand.java, FindCommand.java, InvalidCommand.java, ListCommand.java and Mark.java and update Ui.java, Storage.java, Parser.java and Duke.java
There were methods that were too long which caused deep nesting
in Storage.java and Parser.java.

The Command class was added to reduce the length of methods by
making parser return a specific child class of the Command class,
the required operation is then executed depending on the command
class.